// models/classify/nearest-neighbor/nearest-neighbor-worker.ts
import { Analytic, Log, Statistic } from "@/lib/db";
import { KNNType } from "./nearest-neighbor";

export type KNNAnalysisType = {
    configData: KNNType;
    dataVariables: any[];
    variables: any[];
    addLog: (log: Omit<Log, "id">) => Promise<number>;
    addAnalytic: (analytic: Omit<Analytic, "id">) => Promise<number>;
    addStatistic: (stat: Omit<Statistic, "id">) => Promise<number>;
};

export type KNNFinalResultType = {
    addLog: (log: Omit<Log, "id">) => Promise<number>;
    addAnalytic: (analytic: Omit<Analytic, "id">) => Promise<number>;
    addStatistic: (stat: Omit<Statistic, "id">) => Promise<number>;
};